Remove unnecessary classes, refactor packages
