--- src/test/java/uk/gov/hmcts/reform/rsecheck/DemoUnitTest.java
@@ -1,13 +1,0 @@
-package uk.gov.hmcts.reform.rsecheck;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-class DemoUnitTest {
-
-    @Test
-    void exampleOfTest() {
-        assertTrue(System.currentTimeMillis() > 0, "Example of Unit Test");
-    }
-}
